fix: improve CLI load time
