+++ src/uipath/platform/action_center/_tasks_service.py
@@ -10,8 +10,7 @@ from ..._utils.constants import (
     HEADER_TENANT_ID,
 )
 from ...tracing import traced
-from ..common import FolderContext, UiPathApiConfig, UiPathExecutionContext
-from ..common._base_service import BaseService
+from ..common import BaseService, FolderContext, UiPathApiConfig, UiPathExecutionContext
 from .task_schema import TaskSchema
 from .tasks import Task
